+++ dwds/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 18.0.3-dev
 
+- Fix `getObject` failure on record class.
+
 ## 18.0.2
 
 - Support new DDC temp names for patterns. - [#2042](https://github.com/dart-lang/webdev/pull/2042)